mergeThreeTime
Remain one problem about long click
