+++ app/src/main/java/com/group4sweng/scranplan/MealPlanner/MealTimescaleFragment.java
@@ -27,8 +27,9 @@ import com.google.firebase.firestore.Query;
 import com.google.firebase.firestore.QuerySnapshot;
 import com.group4sweng.scranplan.R;
 import com.group4sweng.scranplan.RecipeInfo.RecipeInfoFragment;
-import com.group4sweng.scranplan.SearchFunctions.BreakfastQueries;
-import com.group4sweng.scranplan.SearchFunctions.BreakfastRecyclerAdapter;
+import com.group4sweng.scranplan.SearchFunctions.MealTimescaleRecyclerAdapter;
+import com.group4sweng.scranplan.SearchFunctions.MealTimescaleQueries;
+import com.group4sweng.scranplan.SearchFunctions.MealTimescaleRecyclerAdapter;
 import com.group4sweng.scranplan.UserInfo.UserInfoPrivate;
 
 import java.util.ArrayList;
@@ -38,10 +39,10 @@ import java.util.Map;
 
 import static androidx.constraintlayout.widget.Constraints.TAG;
 
-public class BreakfastFragment extends Fragment {
+public class MealTimescaleFragment extends Fragment {
 
     UserInfoPrivate user;
-    public BreakfastFragment(UserInfoPrivate userSent){
+    public MealTimescaleFragment(UserInfoPrivate userSent){
         user = userSent;
     }
 
@@ -49,12 +50,22 @@ public class BreakfastFragment extends Fragment {
     private String imageURL;
     private DocumentSnapshot document;
 
-    List<BreakfastRecyclerAdapter.BreakfastRecipePreviewData> dataBreakfast;
+    List<MealTimescaleRecyclerAdapter.MealTimescaleRecipePreviewData> dataBreakfast;
     private DocumentSnapshot lastVisibleBreakfast;
     private boolean isScrollingBreakfast = false;
     private boolean isLastItemReachedBreakfast = false;
 
-    List<BreakfastRecyclerAdapter.BreakfastRecipePreviewData> dataFave;
+    List<MealTimescaleRecyclerAdapter.MealTimescaleRecipePreviewData> dataLunch;
+    private DocumentSnapshot lastVisibleLunch;
+    private boolean isScrollingLunch = false;
+    private boolean isLastItemReachedLunch = false;
+
+    List<MealTimescaleRecyclerAdapter.MealTimescaleRecipePreviewData> dataDinner;
+    private DocumentSnapshot lastVisibleDinner;
+    private boolean isScrollingDinner = false;
+    private boolean isLastItemReachedDinner = false;
+
+    List<MealTimescaleRecyclerAdapter.MealTimescaleRecipePreviewData> dataFave;
     private DocumentSnapshot lastVisibleFave;
     private boolean isScrollingFave = false;
     private boolean isLastItemReachedFave = false;
@@ -89,7 +100,7 @@ public class BreakfastFragment extends Fragment {
         final int scrollViewSize = 5;
 
         if(user != null) {
-            BreakfastQueries horizontalScrollQueries = new BreakfastQueries(user);
+            MealTimescaleQueries horizontalScrollQueries = new MealTimescaleQueries(user);
 
             /* Adding the save view as score but with breakfast as a new query
             /  This has been done in the same manner but as there are too many variables to track
@@ -99,7 +110,7 @@ public class BreakfastFragment extends Fragment {
             recyclerViewBreakfast.setLayoutManager(rBreakfast);
             recyclerViewBreakfast.setLayoutParams(new LinearLayout.LayoutParams(displayMetrics.widthPixels, displayMetrics.heightPixels / scrollViewSize));
             dataBreakfast = new ArrayList<>();
-            final RecyclerView.Adapter rAdapterBreakfast = new BreakfastRecyclerAdapter(BreakfastFragment.this, dataBreakfast);
+            final RecyclerView.Adapter rAdapterBreakfast = new MealTimescaleRecyclerAdapter(MealTimescaleFragment.this, dataBreakfast);
             recyclerViewBreakfast.setAdapter(rAdapterBreakfast);
             final Query queryBreakfast = (Query) horizontalScrollQueries.getQueries().get("breakfast");
             if (queryBreakfast != null) {
@@ -119,7 +130,7 @@ public class BreakfastFragment extends Fragment {
                     public void onComplete(@NonNull Task<QuerySnapshot> task) {
                         if (task.isSuccessful()) {
                             for (DocumentSnapshot document : task.getResult()) {
-                                dataBreakfast.add(new BreakfastRecyclerAdapter.BreakfastRecipePreviewData(
+                                dataBreakfast.add(new MealTimescaleRecyclerAdapter.MealTimescaleRecipePreviewData(
                                         document,
                                         document.getId(),
                                         document.get("Name").toString(),
@@ -160,7 +171,7 @@ public class BreakfastFragment extends Fragment {
                                             public void onComplete(@NonNull Task<QuerySnapshot> t) {
                                                 if (t.isSuccessful()) {
                                                     for (DocumentSnapshot d : t.getResult()) {
-                                                        dataBreakfast.add(new BreakfastRecyclerAdapter.BreakfastRecipePreviewData(
+                                                        dataBreakfast.add(new MealTimescaleRecyclerAdapter.MealTimescaleRecipePreviewData(
                                                                 d,
                                                                 d.getId(),
                                                                 d.get("Name").toString(),
@@ -194,6 +205,213 @@ public class BreakfastFragment extends Fragment {
                 Log.e(TAG, "Breakfast horizontal view added");
             }
 
+             /* Adding the save view as score but with lunch as a new query
+            /  This has been done in the same manner but as there are too many variables to track
+            /  this is not workable in any kind of loop. */
+            final RecyclerView recyclerViewLunch = new RecyclerView(view.getContext());
+            RecyclerView.LayoutManager rType = new LinearLayoutManager(getContext(), LinearLayoutManager.HORIZONTAL, false);
+            recyclerViewLunch.setLayoutManager(rType);
+            recyclerViewLunch.setLayoutParams(new LinearLayout.LayoutParams(displayMetrics.widthPixels, displayMetrics.heightPixels / scrollViewSize));
+            dataLunch = new ArrayList<>();
+            final MealTimescaleRecyclerAdapter rAdapterLunch = new MealTimescaleRecyclerAdapter(MealTimescaleFragment.this, dataLunch);
+            recyclerViewLunch.setAdapter(rAdapterLunch);
+            final Query queryLunch = (Query) horizontalScrollQueries.getQueries().get("lunch");
+            if (queryLunch != null) {
+                Log.e(TAG, "User is searching the following query: " + queryLunch.toString());
+
+                TextView textView = new TextView(view.getContext());
+                String testString = "Lunch";
+                textView.setTextSize(25);
+                textView.setPadding(20, 5, 5, 5);
+                textView.setTextColor(Color.WHITE);
+                textView.setShadowLayer(4, 0, 0, Color.BLACK);
+                textView.setText(testString);
+
+                queryLunch
+                        .get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+                    @Override
+                    public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                        if (task.isSuccessful()) {
+                            for (DocumentSnapshot document : task.getResult()) {
+                                dataLunch.add(new MealTimescaleRecyclerAdapter.MealTimescaleRecipePreviewData(
+                                        document,
+                                        document.getId(),
+                                        document.get("Name").toString(),
+                                        Float.valueOf(document.get("score").toString()),
+                                        document.get("imageURL").toString()
+                                ));
+                            }
+                            rAdapterLunch.notifyDataSetChanged();
+                            if (task.getResult().size() != 0) {
+                                lastVisibleLunch = task.getResult().getDocuments().get(task.getResult().size() - 1);
+                            } else {
+                                isLastItemReachedLunch = true;
+                            }
+
+                            RecyclerView.OnScrollListener onScrollListener = new RecyclerView.OnScrollListener() {
+                                @Override
+                                public void onScrollStateChanged(RecyclerView recyclerView, int newState) {
+                                    super.onScrollStateChanged(recyclerView, newState);
+                                    if (newState == AbsListView.OnScrollListener.SCROLL_STATE_TOUCH_SCROLL) {
+                                        isScrollingLunch = true;
+                                    }
+                                }
+
+                                @Override
+                                public void onScrolled(RecyclerView recyclerView, int dx, int dy) {
+                                    super.onScrolled(recyclerView, dx, dy);
+
+                                    LinearLayoutManager linearLayoutManager = ((LinearLayoutManager) recyclerView.getLayoutManager());
+                                    int firstVisibleItemPosition = linearLayoutManager.findFirstVisibleItemPosition();
+                                    int visibleItemCount = linearLayoutManager.getChildCount();
+                                    int totalItemCount = linearLayoutManager.getItemCount();
+
+                                    if (isScrollingLunch && (firstVisibleItemPosition + visibleItemCount == totalItemCount) && !isLastItemReachedBreakfast) {
+                                        isScrollingLunch = false;
+                                        Query nextQuery = queryLunch.startAfter(lastVisibleLunch);
+                                        nextQuery.get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+                                            @Override
+                                            public void onComplete(@NonNull Task<QuerySnapshot> t) {
+                                                if (t.isSuccessful()) {
+                                                    for (DocumentSnapshot d : t.getResult()) {
+                                                        dataBreakfast.add(new MealTimescaleRecyclerAdapter.MealTimescaleRecipePreviewData(
+                                                                d,
+                                                                d.getId(),
+                                                                d.get("Name").toString(),
+                                                                Float.valueOf(d.get("score").toString()),
+                                                                d.get("imageURL").toString()
+                                                        ));
+                                                    }
+                                                    if (isLastItemReachedLunch) {
+                                                        // Add end here
+                                                    }
+                                                    rAdapterLunch.notifyDataSetChanged();
+                                                    if (t.getResult().size() != 0) {
+                                                        lastVisibleLunch = t.getResult().getDocuments().get(t.getResult().size() - 1);
+                                                    }
+
+                                                    if (t.getResult().size() < 10) {
+                                                        isLastItemReachedLunch = true;
+                                                    }
+                                                }
+                                            }
+                                        });
+                                    }
+                                }
+                            };
+                            recyclerViewLunch.addOnScrollListener(onScrollListener);
+                        }
+                    }
+                });
+                topLayout.addView(textView);
+                topLayout.addView(recyclerViewLunch);
+                Log.e(TAG, "Breakfast horizontal view added");
+            }
+
+            /* Adding the save view as score but with dinner as a new query
+            /  This has been one in the same manner but as there are too many variables to track
+            /  this is not workable in any kind of loop. */
+            final RecyclerView recyclerViewDinner = new RecyclerView(view.getContext());
+            RecyclerView.LayoutManager rDinner = new LinearLayoutManager(getContext(), LinearLayoutManager.HORIZONTAL, false);
+            recyclerViewDinner.setLayoutManager(rDinner);
+            recyclerViewDinner.setLayoutParams(new LinearLayout.LayoutParams(displayMetrics.widthPixels, displayMetrics.heightPixels / scrollViewSize));
+            dataDinner = new ArrayList<>();
+            final RecyclerView.Adapter rAdapterDinner = new MealTimescaleRecyclerAdapter(MealTimescaleFragment.this, dataDinner);
+            recyclerViewDinner.setAdapter(rAdapterDinner);
+            final Query queryDinner = (Query) horizontalScrollQueries.getQueries().get("dinner");
+            if (queryDinner != null) {
+                Log.e(TAG, "User is searching the following query: " + queryDinner.toString());
+
+                TextView textView = new TextView(view.getContext());
+                String testString = "Dinner";
+                textView.setTextSize(25);
+                textView.setPadding(20, 5, 5, 5);
+                textView.setTextColor(Color.WHITE);
+                textView.setShadowLayer(4, 0, 0, Color.BLACK);
+                textView.setText(testString);
+
+                queryDinner
+                        .get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+                    @Override
+                    public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                        if (task.isSuccessful()) {
+                            for (DocumentSnapshot document : task.getResult()) {
+                                dataDinner.add(new MealTimescaleRecyclerAdapter.MealTimescaleRecipePreviewData(
+                                        document,
+                                        document.getId(),
+                                        document.get("Name").toString(),
+                                        Float.valueOf(document.get("score").toString()),
+                                        document.get("imageURL").toString()
+                                ));
+                            }
+                            rAdapterDinner.notifyDataSetChanged();
+                            if (task.getResult().size() != 0) {
+                                lastVisibleDinner = task.getResult().getDocuments().get(task.getResult().size() - 1);
+                            } else {
+                                isLastItemReachedDinner = true;
+                            }
+
+                            RecyclerView.OnScrollListener onScrollListener = new RecyclerView.OnScrollListener() {
+                                @Override
+                                public void onScrollStateChanged(RecyclerView recyclerView, int newState) {
+                                    super.onScrollStateChanged(recyclerView, newState);
+                                    if (newState == AbsListView.OnScrollListener.SCROLL_STATE_TOUCH_SCROLL) {
+                                        isScrollingDinner = true;
+                                    }
+                                }
+
+                                @Override
+                                public void onScrolled(RecyclerView recyclerView, int dx, int dy) {
+                                    super.onScrolled(recyclerView, dx, dy);
+
+                                    LinearLayoutManager linearLayoutManager = ((LinearLayoutManager) recyclerView.getLayoutManager());
+                                    int firstVisibleItemPosition = linearLayoutManager.findFirstVisibleItemPosition();
+                                    int visibleItemCount = linearLayoutManager.getChildCount();
+                                    int totalItemCount = linearLayoutManager.getItemCount();
+
+                                    if (isScrollingDinner && (firstVisibleItemPosition + visibleItemCount == totalItemCount) && !isLastItemReachedDinner) {
+                                        isScrollingDinner = false;
+                                        Query nextQuery = queryDinner.startAfter(lastVisibleDinner);
+                                        nextQuery.get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+                                            @Override
+                                            public void onComplete(@NonNull Task<QuerySnapshot> t) {
+                                                if (t.isSuccessful()) {
+                                                    for (DocumentSnapshot d : t.getResult()) {
+                                                        dataDinner.add(new MealTimescaleRecyclerAdapter.MealTimescaleRecipePreviewData(
+                                                                d,
+                                                                d.getId(),
+                                                                d.get("Name").toString(),
+                                                                Float.valueOf(d.get("score").toString()),
+                                                                d.get("imageURL").toString()
+                                                        ));
+                                                    }
+                                                    if (isLastItemReachedDinner) {
+                                                        // Add end here
+                                                    }
+                                                    rAdapterDinner.notifyDataSetChanged();
+                                                    if (t.getResult().size() != 0) {
+                                                        lastVisibleDinner = t.getResult().getDocuments().get(t.getResult().size() - 1);
+                                                    }
+
+                                                    if (t.getResult().size() < 5) {
+                                                        isLastItemReachedDinner = true;
+                                                    }
+                                                }
+                                            }
+                                        });
+                                    }
+                                }
+                            };
+                            recyclerViewDinner.addOnScrollListener(onScrollListener);
+                        }
+                    }
+                });
+                topLayout.addView(textView);
+                topLayout.addView(recyclerViewDinner);
+                Log.e(TAG, "Dinner horizontal view added");
+            }
+
+
             /* Adding the save view as score but with user favourite recipes as a new query
             /  This has been done in the same manner but as there are too many variables to track
             /  this is not workable in any kind of loop. */
@@ -202,7 +420,7 @@ public class BreakfastFragment extends Fragment {
             recyclerViewFave.setLayoutManager(rManagerFave);
             recyclerViewFave.setLayoutParams(new LinearLayout.LayoutParams(displayMetrics.widthPixels, displayMetrics.heightPixels / scrollViewSize));
             dataFave = new ArrayList<>();
-            final RecyclerView.Adapter rAdapterFave = new BreakfastRecyclerAdapter(BreakfastFragment.this, dataFave);
+            final RecyclerView.Adapter rAdapterFave = new MealTimescaleRecyclerAdapter(MealTimescaleFragment.this, dataFave);
             recyclerViewFave.setAdapter(rAdapterFave);
             final Query queryFave = (Query) horizontalScrollQueries.getQueries().get("favourite");
             if (queryFave != null) {
@@ -222,7 +440,7 @@ public class BreakfastFragment extends Fragment {
                     public void onComplete(@NonNull Task<QuerySnapshot> task) {
                         if (task.isSuccessful()) {
                             for (DocumentSnapshot document : task.getResult()) {
-                                dataFave.add(new BreakfastRecyclerAdapter.BreakfastRecipePreviewData(
+                                dataFave.add(new MealTimescaleRecyclerAdapter.MealTimescaleRecipePreviewData(
                                         document,
                                         document.getId(),
                                         document.get("Name").toString(),
@@ -263,7 +481,7 @@ public class BreakfastFragment extends Fragment {
                                             public void onComplete(@NonNull Task<QuerySnapshot> t) {
                                                 if (t.isSuccessful()) {
                                                     for (DocumentSnapshot d : t.getResult()) {
-                                                        dataFave.add(new BreakfastRecyclerAdapter.BreakfastRecipePreviewData(
+                                                        dataFave.add(new MealTimescaleRecyclerAdapter.MealTimescaleRecipePreviewData(
                                                                 d,
                                                                 d.getId(),
                                                                 d.get("Name").toString(),
@@ -331,9 +549,22 @@ public class BreakfastFragment extends Fragment {
             mBundle.putString("recipeDescription", document.get("Description").toString());
             mBundle.putString("chefName", document.get("Chef").toString());
             mBundle.putBoolean("planner", planner);
+            mBundle.putBoolean("canFreeze", document.getBoolean("freezer"));
+            mBundle.putString("peopleServes", document.get("serves").toString());
+            mBundle.putString("fridgeDays", document.get("fridge").toString());
+            mBundle.putString("reheat", document.get("reheat").toString());
+            mBundle.putBoolean("noEggs", document.getBoolean("noEggs"));
+            mBundle.putBoolean("noMilk", document.getBoolean("noMilk"));
+            mBundle.putBoolean("noNuts", document.getBoolean("noNuts"));
+            mBundle.putBoolean("noShellfish", document.getBoolean("noShellfish"));
+            mBundle.putBoolean("noSoy", document.getBoolean("noSoy"));
+            mBundle.putBoolean("noWheat", document.getBoolean("noWheat"));
+            mBundle.putBoolean("pescatarian", document.getBoolean("pescatarian"));
+            mBundle.putBoolean("vegan", document.getBoolean("vegan"));
+            mBundle.putBoolean("vegetarian", document.getBoolean("vegetarian"));
 
             ArrayList faves = (ArrayList) document.get("favourite");
-            mBundle.putBoolean("isFav", faves.contains(user.getUID()));
+            mBundle.putBoolean("isFav", faves.contains(user.getUID().hashCode()));
 
 
             RecipeInfoFragment recipeDialogFragment = new RecipeInfoFragment();
